update changes from andrew's code review
